fix: move constants
