+++ packages/snap/src/constants/index.ts
@@ -1,0 +1,8 @@
+export const availableVCStores = ['snap', 'ceramic'] as const;
+
+export const availableMethods = ['did:ethr', 'did:key'] as const;
+
+export const didCoinTypeMappping: Record<string, number> = {
+  'did:ethr': 60,
+  'did:key': 60,
+};
