--- packages/snap/src/did/didMethods.ts
@@ -1,1 +1,0 @@
-export const availableMethods = ['did:ethr', 'did:key'] as const;
