+++ packages/snap/src/rpc/did/getAvailableMethods.ts
@@ -1,4 +1,4 @@
-import { availableMethods } from '../../did/didMethods';
+import { availableMethods } from '../../constants/index';
 
 export function getAvailableMethods(): string[] {
   return availableMethods.map((key) => key);
