+++ packages/snap/src/rpc/vcStore/getAvailableVCStores.ts
@@ -1,4 +1,4 @@
-import { availableVCStores } from '../../veramo/plugins/availableVCStores';
+import { availableVCStores } from '../../constants/index';
 
 export function getAvailableVCStores(): string[] {
   return availableVCStores.map((key) => key);
