--- packages/snap/src/veramo/plugins/availableVCStores.ts
@@ -1,1 +1,0 @@
-export const availableVCStores = ['snap', 'ceramic'] as const;
